separate queue-table-by-status into its own file, remove location-related code from it
