+++ packages/esm-service-queues-app/src/root.component.tsx
@@ -5,7 +5,7 @@ import AppointmentsTable from './queue-patient-linelists/scheduled-appointments-
 import Home from './home.component';
 import ServicesTable from './queue-patient-linelists/queue-services-table.component';
 import QueueScreen from './queue-screen/queue-screen.component';
-import { QueueTableByStatus } from './queue-table/queue-table.component';
+import { QueueTableByStatus } from './queue-table/queue-table-by-status.component';
 
 const swrConfiguration = {
   // Maximum number of retries when the backend returns an error
