+++ packages/esm-service-queues-app/translations/en.json
@@ -24,6 +24,7 @@
   "age": "Age",
   "alistOfClients": "A list of clients waiting for ",
   "all": "All",
+  "allQueues": "All Queues",
   "and": "And",
   "anotherVisitType": "Start another visit type",
   "any": "Any",
